Mac: clean-up of incognito badge code and float -> CGFloat.
BUG=none
TEST=everything still builds and works

Review URL: http://codereview.chromium.org/669277

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@40906 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/browser/cocoa/view_resizer.h
@@ -16,7 +16,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 // to perform the resize.  This allows the parent to do any re-layout that may
 // become necessary due to the resize.
 @protocol ViewResizer <NSObject>
-- (void)resizeView:(NSView*)view newHeight:(float)height;
+- (void)resizeView:(NSView*)view newHeight:(CGFloat)height;
 
 @optional
 // Optional method called when an animation is beginning or ending.  Resize
