+++ chrome/browser/cocoa/view_resizer_pong.h
@@ -12,11 +12,11 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 @interface ViewResizerPong : NSObject<ViewResizer> {
  @private
-  float height_;
+  CGFloat height_;
 }
-@property float height;
+@property CGFloat height;
 
-- (void)resizeView:(NSView*)view newHeight:(float)height;
+- (void)resizeView:(NSView*)view newHeight:(CGFloat)height;
 @end
 
 #endif  // CHROME_BROWSER_COCOA_VIEW_RESIZER_PONG_H_
