+++ chrome/browser/cocoa/view_resizer_pong.mm
@@ -11,7 +11,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 @synthesize height = height_;
 
-- (void)resizeView:(NSView*)view newHeight:(float)height {
+- (void)resizeView:(NSView*)view newHeight:(CGFloat)height {
   [self setHeight:height];
 
   // Set the view's height and width, in case it uses that as important state.
